Remove GBA-specific SWI code.

+++ BeeARM-tests/testenv.cpp
@@ -75,7 +75,7 @@ class TestInterface : public BeeARMInterface
 	return;
     }
 
-    void softwareinterrupt(uint8_t val)
+    void softwareinterrupt(uint32_t val)
     {
 	return;
     }
